# doc-api/api-doc.json
{
    "openapi": "3.0.4",
    "info": {
      "title": "CuroTec.API",
      "version": "1.0"
    },
    "paths": {
      "/vehicles": {
        "get": {
          "tags": [
            "CuroTec.API"
          ],
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        },
        "post": {
          "tags": [
            "CuroTec.API"
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VehicleDto"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        }
      },
      "/vehicles/{id}": {
        "get": {
          "tags": [
            "CuroTec.API"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        },
        "put": {
          "tags": [
            "CuroTec.API"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VehicleDto"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        },
        "delete": {
          "tags": [
            "CuroTec.API"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        }
      },
      "/vehicles/type/{type}": {
        "get": {
          "tags": [
            "CuroTec.API"
          ],
          "parameters": [
            {
              "name": "type",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        }
      },
      "/vehicles/color": {
        "get": {
          "tags": [
            "CuroTec.API"
          ],
          "parameters": [
            {
              "name": "color",
              "in": "query",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        }
      }
    },
    "components": {
      "schemas": {
        "VehicleDto": {
          "type": "object",
          "properties": {
            "vehicleType": {
              "$ref": "#/components/schemas/VehicleType"
            },
            "color": {
              "type": "string",
              "nullable": true
            }
          },
          "additionalProperties": false
        },
        "VehicleType": {
          "enum": [
            0,
            1,
            2,
            3
          ],
          "type": "integer",
          "format": "int32"
        }
      }
    }
  }